+++ src/platform/arcan/video.c
@@ -411,8 +411,11 @@ int platform_video_cardhandle(int cardn, int* method, size_t* msz, uint8_t** dbu
 	return -1;
 }
 
-void platform_event_translation(int devid, int action, const char** names)
+bool platform_event_translation(int devid,
+	int action, const char** names, const char** err)
 {
+	*err = "Not Supported";
+	return false;
 }
 
 int platform_event_device_request(int space, const char* path)