+++ src/platform/stub/event.c
@@ -28,8 +28,11 @@ void platform_event_analogall(bool enable, bool mouse)
 {
 }
 
-void platform_event_translation(int devid, int action, const char** names)
+bool platform_event_translation(
+	int devid, int action, const char** names, const char** err)
 {
+	*err = "Unsupported";
+	return false;
 }
 
 int platform_event_device_request(int space, const char* path)