+++ src/platform/syscons/event.c
@@ -251,8 +251,11 @@ static struct bsdkey decode_tok(char* tok)
 	return res;
 }
 
-void platform_event_translation(int devid, int action, const char** names)
+bool platform_event_translation(
+	int devid, int action, const char** names, const char** err)
 {
+	*err = "Unsupported";
+	return false;
 }
 
 int platform_event_device_request(int space, const char* path)